+++ backend/chalice/api_server/.chalice/config.json
@@ -1,0 +1,27 @@
+{
+  "stages": {
+    "test": {
+      "lambda_memory_size": 256,
+      "lambda_timeout": 30,
+      "security_group_ids": [
+        ""
+      ],
+      "subnet_ids": [],
+      "environment_variables": {
+        "APP_NAME": "corpora-api",
+        "DEPLOYMENT_STAGE": "test",
+        "VERSION": "d845c3e"
+      },
+      "tags": {
+        "created-from": "corpora browser",
+        "env": "test",
+        "service": "corpora-api"
+      },
+      "iam_policy_file": "policy-test.json",
+      "api_gateway_stage": "test"
+    }
+  },
+  "version": "2.0",
+  "app_name": "corpora-api",
+  "autogen_policy": false
+}
