+++ backend/corpora/common/utils/db_session.py
@@ -6,7 +6,6 @@ from sqlalchemy.orm import sessionmaker, session
 
 from .exceptions import CorporaException
 from ..corpora_config import CorporaDbConfig
-from ..corpora_orm import Base, DbDatasetProcessingStatus
 
 logger = logging.getLogger(__name__)
 
@@ -25,24 +24,6 @@ class DBSessionMaker:
 
     def session(self, **kwargs) -> session.Session:
         return self._session_make(**kwargs)
-
-
-def clone(model: Base, primary_key: dict = None, **kwargs) -> Base:
-    """Clone an arbitrary sqlalchemy model object with new primary keys.
-    https://stackoverflow.com/questions/28871406/how-to-clone-a-sqlalchemy-db-object-with-new-primary-key
-
-    :param model: The SQLAlchemy model to clone
-    :param primary_key: The new primary key values. If None a new primary key is generated
-    :param kwargs: Updates the columns in the cloned model.
-    :return: a clone of the original model with any kwargs passed in and the new primary key.
-    """
-    table = model.__table__
-    non_pk_columns = [key for key in table.columns.keys() if key not in table.primary_key]
-    data = {column: getattr(model, column) for column in non_pk_columns}
-    data.update(kwargs)
-    if primary_key:
-        data.update(primary_key)
-    return model.__class__(**data)
 
 
 @contextmanager
@@ -65,23 +46,3 @@ def db_session_manager(**kwargs):
         raise CorporaException(msg)
     finally:
         session.close()
-
-
-def processing_status_updater(session, uuid: str, updates: dict):
-    session.query(DbDatasetProcessingStatus).filter(DbDatasetProcessingStatus.id == uuid).update(updates)
-    session.commit()
-    logger.debug("updating status", updates)
-
-
-from flask import g
-from functools import wraps
-
-
-def dbconnect(func):
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        with db_session_manager() as session:
-            g.db_session = session
-            return func(*args, **kwargs)
-
-    return wrapper
