+++ backend/corpora/lambdas/api/v1/collection.py
@@ -6,7 +6,7 @@ from flask import make_response, jsonify, g
 from ....common.corpora_orm import DbCollection, CollectionVisibility
 from ....common.entities import Collection
 from ....common.utils.exceptions import ForbiddenHTTPException, ConflictException
-from ....common.utils.db_session import dbconnect
+from ....api_server.db import dbconnect
 
 
 @dbconnect
