+++ backend/corpora/lambdas/api/v1/collection_uuid/gene_set.py
@@ -4,7 +4,7 @@ from sqlalchemy.exc import IntegrityError
 from .....common.corpora_orm import CollectionVisibility
 from .....common.entities import Collection
 from .....common.entities.geneset import Geneset
-from .....common.utils.db_session import dbconnect
+from .....api_server.db import dbconnect
 from .....common.utils.exceptions import ForbiddenHTTPException, InvalidParametersHTTPException
