+++ backend/corpora/lambdas/upload_failures/upload.py
@@ -1,6 +1,6 @@
 from backend.corpora.common.entities import Dataset
 from backend.corpora.common.corpora_orm import UploadStatus, DbDatasetProcessingStatus
-from backend.corpora.common.utils.db_session import processing_status_updater
+from backend.corpora.common.utils.db_helpers import processing_status_updater
 from backend.corpora.common.utils.db_session import db_session_manager
